crypto-client-api/requirements.txt
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
Flask==1.1.1
flask-cors==3.0.8
idna==2.8
numpy==1.18.1
oauthlib==3.1.0
orjson==3.3.0
protobuf==3.13.0
py-healthcheck==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyMySQL==0.9.3
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.10.5
pytz==2019.3
rsa==4.0
simplejson==3.17.0
six==1.13.0
sympy==1.5.1
Werkzeug==0.16.0
wrapt==1.12.1
urllib3==1.25.8
